requirements.txt
alembic==1.5.7
aniso8601==9.0.1
asyncio==3.4.3
certifi==2020.12.5
click==7.1.2
Flask==0.12.2
Flask-Cors==3.0.10
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
flask-serialize==1.5.1
Flask-SQLAlchemy==2.3.2
greenlet==1.0.0
gunicorn==20.0.4
importlib-metadata==3.7.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==2.14.0
marshmallow-sqlalchemy==0.13.2
numpy==1.20.1
pandas==1.1.5
permissive-dict==1.0.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.2
SQLAlchemy-serializer==1.3.4.4
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
